Create app with models and migrations
