+++ api/features/import_export/models.py
@@ -1,0 +1,65 @@
+from django.db import models
+
+from features.import_export.constants import (
+    FEATURE_IMPORT_STATUSES,
+    FEATURE_IMPORT_STRATEGIES,
+    MAX_FEATURE_EXPORT_SIZE,
+    MAX_FEATURE_IMPORT_SIZE,
+    PROCESSING,
+)
+
+
+class FeatureExport(models.Model):
+    """
+    Stores the representation of an environment's export of
+    features between the request for the export and the
+    ultimate download. Records are deleted automatically after
+    a waiting period.
+    """
+
+    # The environment the export came from.
+    environment = models.ForeignKey(
+        "environments.Environment",
+        related_name="feature_imports",
+        on_delete=models.CASCADE,
+        swappable=True,
+    )
+
+    # This is a JSON string of data used for file download
+    # once the task has completed assembly.
+    data = models.CharField(max_length=MAX_FEATURE_EXPORT_SIZE)
+    created_at = models.DateTimeField(auto_now_add=True)
+
+
+class FeatureImport(models.Model):
+    """
+    Stores the representation of an environment's import of
+    features between upload of a previously exported featureset
+    and the processing of the import. Records are deleted
+    automatically after a waiting period.
+    """
+
+    # The environment the features are being imported to.
+    environment = models.ForeignKey(
+        "environments.Environment",
+        related_name="feature_exports",
+        on_delete=models.CASCADE,
+        swappable=True,
+    )
+    strategy = models.CharField(
+        choices=FEATURE_IMPORT_STRATEGIES,
+        max_length=50,
+        blank=False,
+        null=False,
+    )
+    status = models.CharField(
+        choices=FEATURE_IMPORT_STATUSES,
+        max_length=50,
+        blank=False,
+        null=False,
+        default=PROCESSING,
+    )
+
+    # This is a JSON string of data generated by the export.
+    data = models.CharField(max_length=MAX_FEATURE_IMPORT_SIZE)
+    created_at = models.DateTimeField(auto_now_add=True)
